+++ GANDLF/compute/training_loop.py
@@ -15,7 +15,10 @@ from GANDLF.utils import (
     load_model,
     version_check,
     write_training_patches,
+    print_model_summary,
+    get_ground_truths_and_predictions_tensor,
 )
+from GANDLF.metrics import overall_stats
 from GANDLF.logger import Logger
 from .step import step
 from .forward_pass import validate_network
@@ -68,6 +71,12 @@ def train_network(model, train_dataloader, optimizer, params):
         if params["verbose"]:
             print("Using Automatic mixed precision", flush=True)
 
+    # get ground truths
+    if params["problem_type"] == "classification":
+        (
+            ground_truth_array,
+            predictions_array,
+        ) = get_ground_truths_and_predictions_tensor(params, "training_data")
     # Set the model to train
     model.train()
     for batch_idx, (subject) in enumerate(
@@ -103,7 +112,15 @@ def train_network(model, train_dataloader, optimizer, params):
             params["subject_spacing"] = subject["spacing"]
         else:
             params["subject_spacing"] = None
-        loss, calculated_metrics, _ = step(model, image, label, params)
+        loss, calculated_metrics, output, _ = step(model, image, label, params)
+        # store predictions for classification
+        if params["problem_type"] == "classification":
+            predictions_array[
+                batch_idx
+                * params["batch_size"] : (batch_idx + 1)
+                * params["batch_size"]
+            ] = (torch.argmax(output[0], 0).cpu().item())
+
         nan_loss = torch.isnan(loss)
         second_order = (
             hasattr(optimizer, "is_second_order") and optimizer.is_second_order
@@ -157,7 +174,7 @@ def train_network(model, train_dataloader, optimizer, params):
                 (batch_idx + 1) < len(train_dataloader)
             ):
                 print(
-                    "\nHalf-Epoch Average Train loss : ",
+                    "\nHalf-Epoch Average train loss : ",
                     total_epoch_train_loss / (batch_idx + 1),
                 )
                 for metric in params["metrics"]:
@@ -168,12 +185,18 @@ def train_network(model, train_dataloader, optimizer, params):
                     else:
                         to_print = total_epoch_train_metric[metric] / (batch_idx + 1)
                     print(
-                        "Half-Epoch Average Train " + metric + " : ",
+                        "Half-Epoch Average train " + metric + " : ",
                         to_print,
                     )
 
     average_epoch_train_loss = total_epoch_train_loss / len(train_dataloader)
-    print("     Epoch Final   Train loss : ", average_epoch_train_loss)
+    print("     Epoch Final   train loss : ", average_epoch_train_loss)
+
+    # get overall stats for classification
+    if params["problem_type"] == "classification":
+        average_epoch_train_metric = overall_stats(
+            predictions_array, ground_truth_array, params
+        )
     for metric in params["metrics"]:
         if isinstance(total_epoch_train_metric[metric], np.ndarray):
             to_print = (
@@ -182,8 +205,9 @@ def train_network(model, train_dataloader, optimizer, params):
         else:
             to_print = total_epoch_train_metric[metric] / len(train_dataloader)
         average_epoch_train_metric[metric] = to_print
+    for metric in average_epoch_train_metric.keys():
         print(
-            "     Epoch Final   Train " + metric + " : ",
+            "     Epoch Final   train " + metric + " : ",
             average_epoch_train_metric[metric],
         )
 
@@ -235,6 +259,15 @@ def training_loop(
         scheduler,
         params,
     ) = create_pytorch_objects(params, training_data, validation_data, device)
+
+    if params["model"]["print_summary"]:
+        print_model_summary(
+            model,
+            params["batch_size"],
+            params["model"]["num_channels"],
+            params["patch_size"],
+            params["device"],
+        )
 
     if testingDataDefined:
         test_dataloader = get_testing_loader(params)
@@ -453,9 +486,7 @@ def training_loop(
         onnx_export = True
         if params["model"]["architecture"] in ["sdnet", "brain_age"]:
             onnx_export = False
-        elif (
-            "onnx_export" in params["model"] and params["model"]["onnx_export"] == False
-        ):
+        elif "onnx_export" in params["model"] and not (params["model"]["onnx_export"]):
             onnx_export = False
 
         if onnx_export: