+++ GANDLF/data/post_process/tensor.py
@@ -1,31 +1,26 @@
-import torch
 import numpy as np
+from GANDLF.utils.generic import get_array_from_image_or_tensor
 
 
 def get_mapped_label(input_tensor, params):
     """
     This function maps the input label to the output label.
     Args:
-        input_tensor (torch.Tensor): The input label.
+        input_tensor (Union[torch.Tensor, sitk.Image]): The input label.
         params (dict): The parameters dict.
 
     Returns:
-        torch.Tensor: The output image after morphological operations.
+        np.ndarray: The output image after morphological operations.
     """
-    input_arr = input_tensor
-    if torch.is_tensor(input_tensor):
-        input_arr = input_tensor.numpy()
-
+    input_image_array = get_array_from_image_or_tensor(input_tensor)
     if "data_postprocessing" not in params:
-        return input_arr
+        return input_image_array
     if "mapping" not in params["data_postprocessing"]:
-        return input_arr
+        return input_image_array
 
-    mapping = params["data_postprocessing"]["mapping"]
+    output = np.zeros(input_image_array.shape)
 
-    output = np.zeros(input_arr.shape)
+    for key, value in params["data_postprocessing"]["mapping"].items():
+        output[input_tensor == key] = value
 
-    for key, value in mapping.items():
-        output[input_arr == key] = value
-
-    return torch.from_numpy(output)
+    return output