+++ GANDLF/utils/generic.py
@@ -1,5 +1,7 @@
 import os, datetime, sys
 import numpy as np
+import torch
+import SimpleITK as sitk
 
 
 def checkPatchDivisibility(patch_size, number=16):
@@ -32,18 +34,6 @@ def checkPatchDivisibility(patch_size, number=16):
     if (unique.shape[0] == 1) and (unique[0] <= number):
         return False
     return True
-
-
-def fix_paths(cwd):
-    """
-    This function takes the current working directory of the script (which is required for VIPS) and sets up all the paths correctly
-
-    Args:
-        cwd (str): The current working directory.
-    """
-    if os.name == "nt":  # proceed for windows
-        vipshome = os.path.join(cwd, "vips/vips-dev-8.10/bin")
-        os.environ["PATH"] = vipshome + ";" + os.environ["PATH"]
 
 
 def get_date_time():
@@ -162,3 +152,21 @@ def getBase2(num):
         num = num / 2
         base = base + 1
     return base
+
+
+def get_array_from_image_or_tensor(input_tensor_or_image):
+    """
+    This function returns the numpy array from a tensor or image.
+    Args:
+        input_tensor_or_image (torch.Tensor or sitk.Image): The input tensor or image.
+    Returns:
+        numpy.array: The numpy array from the tensor or image.
+    """
+    if isinstance(input_tensor_or_image, torch.Tensor):
+        return input_tensor_or_image.detach().numpy()
+    elif isinstance(input_tensor_or_image, sitk.Image):
+        return sitk.GetArrayFromImage(input_tensor_or_image)
+    elif isinstance(input_tensor_or_image, np.ndarray):
+        return input_tensor_or_image
+    else:
+        raise ValueError("Input must be a torch.Tensor or sitk.Image or np.ndarray")